move button handling from TrackerInput to TrackerButtons
