+++ librefit-web/src/lib/components/TrackerInput.svelte
@@ -1,19 +1,12 @@
 <script>
-	import { createEventDispatcher } from 'svelte';
-	import Trash from '$lib/assets/icons/trash.svg?component';
-	import AddKcal from '$lib/assets/icons/hamburger-plus.svg?component';
-	import AddWeight from '$lib/assets/icons/plus.svg?component';
-	import Add from '$lib/assets/icons/plus.svg?component';
-	import Edit from '$lib/assets/icons/pencil.svg?component';
-	import Check from '$lib/assets/icons/check.svg?component';
-	import CancelDelete from '$lib/assets/icons/trash-off.svg?component';
-	import CancelEdit from '$lib/assets/icons/pencil-off.svg?component';
+	import {createEventDispatcher} from 'svelte';
 	import {getDaytimeFoodCategory} from '$lib/date.js';
+	import TrackerButtons from './TrackerButtons.svelte';
 
 	export let value, dateStr, sequence;
 	export let existing = false;
 	export let disabled = false;
-	export let quickadd = false;
+	export let compact = false;
 
 	/** @type Array<FoodCategory> */
 	export let categories;
@@ -25,43 +18,21 @@
 
 	const dispatch = createEventDispatcher();
 
-	let editing = false;
-
 	let previous;
 	let changeAction;
 
-	let btnAdd, btnConfirm, btnCancel;
-
-	const add = () => {
-		btnAdd.disabled = true;
-
+	const add = (e) => {
 		dispatch('add', {
 			sequence: sequence,
 			dateStr: dateStr,
 			value: value,
 			category: category,
-			callback: () => { btnAdd.disabled = false }
+			callback: () => { e.detail.callback() }
 		});
-	};
-
-	const change = (action) => {
-		return () => {
-			disabled = false;
-			editing = true;
-
-			changeAction = action;
-
-			if (action === 'update') {
-				previous = { category, value };
-			}
-		};
 	};
 
 	const update = (e) => {
 		e.preventDefault();
-
-		btnConfirm.disabled = true;
-		btnCancel.disabled = true;
 
 		if (value !== previous.value || category !== previous.category) {
 			dispatch('update', {
@@ -69,7 +40,7 @@
 				dateStr: dateStr,
 				value: value,
 				category: category,
-				callback: postAction
+				callback: () => e.detail.callback()
 			});
 		} else {
 			postAction();
@@ -79,39 +50,13 @@
 	const remove = (e) => {
 		e.preventDefault();
 
-		btnConfirm.disabled = true;
-		btnCancel.disabled = true;
-
 		dispatch('remove', {
 			sequence: sequence,
 			dateStr: dateStr,
-			target: btnConfirm,
-			callback: postAction
+			target: e.target,
+			callback: () => e.detail.callback()
 		});
 	};
-
-	const discard = () => {
-		disabled = true;
-		editing = false;
-
-		if (changeAction === 'update') {
-			value = previous.value;
-			category = previous.category;
-		}
-	};
-
-	const postAction = (error) => {
-		if (editing) {
-			btnConfirm.disabled = false;
-			btnCancel.disabled = false;
-		}
-
-		if (error) discard();
-		else {
-			disabled = true;
-			editing = false;
-		}
-	}
 </script>
 
 <style>
@@ -133,74 +78,29 @@
 			</select>
 			{/if}
 		</div>
-		{#if quickadd}
-		<div>
-			<button aria-label="add" bind:this={btnAdd} class="btn-icon variant-filled-primary" on:click|preventDefault={add}>
-				<span>
-					{#if unit === 'kcal'}
-						<AddKcal/>
-					{:else if unit === 'kg'}
-						<AddWeight/>
-					{:else}
-						<Add/>
-					{/if}
-				</span>
-			</button>
-		</div>
+		{#if compact}
+		<TrackerButtons {unit} {existing}
+				on:add={add}
+				on:update={update}
+				on:remove={remove}
+				bind:previous={previous}
+				bind:changeAction={changeAction}
+				bind:disabled={disabled}
+				bind:value
+				bind:category
+		/>
 		{/if}
 	</div>
-	{#if !quickadd}
-	<div class="flex flex-row gap-1 justify-end">
-		<div>
-			{#if !existing}
-			<div>
-				<button aria-label="add" bind:this={btnAdd} class="btn-icon variant-filled-primary" on:click|preventDefault={add}>
-					<span>
-					{#if unit === 'kcal'}
-						<AddKcal/>
-					{:else if unit === 'kg'}
-						<AddWeight/>
-					{:else}
-						<Add/>
-					{/if}
-					</span>
-				</button>
-			</div>
-			{:else}
-			{#if !editing}
-			<button aria-label="edit" class="btn-icon variant-filled-secondary" on:click|preventDefault={change('update')}>
-				<span>
-					<Edit/>
-				</span>
-			</button>
-			<button aria-label="delete" class="btn-icon variant-filled" on:click|preventDefault={change('delete')}>
-				<span>
-					<Trash/>
-				</span>
-			</button>
-			{:else}
-			<button aria-label="confirm" bind:this={btnConfirm}
-				class="btn-icon variant-ghost-primary"
-				on:click={changeAction === 'update' ? update : remove}
-			>
-				<span>
-					<Check/>
-				</span>
-			</button>
-			<button aria-label="discard" bind:this={btnCancel} class="btn-icon variant-ghost-error" on:click|preventDefault={discard}>
-				<span>
-					{#if changeAction === 'update'}
-						<CancelEdit/>
-					{:else if changeAction === 'delete'}
-						<CancelDelete/>
-					{:else}
-						X
-					{/if}
-				</span>
-			</button>
-			{/if}
-		{/if}
-		</div>
-	</div>
+	{#if !compact}
+		<TrackerButtons {unit} {existing}
+						on:add={add}
+						on:update={update}
+						on:remove={remove}
+						bind:previous={previous}
+						bind:changeAction={changeAction}
+						bind:disabled={disabled}
+						bind:value
+						bind:category
+		/>
 	{/if}
 </div>
